Contact excel upload (WIP)

+++ modules/Core/Database/Migrations/2020_02_05_182216_create_organization_configurations_table.php
@@ -17,7 +17,7 @@ class CreateOrganizationConfigurationsTable extends Migration
             $table->bigIncrements('id');
 
             $table->unsignedBigInteger('org_id');
-            $table->integer('configuration_id');
+            $table->unsignedBigInteger('configuration_id');
             $table->text('value')->nullable();
 
             //Audit Log Fields
